refactor(minesweeper): BoardCellsCreator 와 구현체를 strategy 패키지로 이동
