+++ src/main/kotlin/tdd/minesweeper/domain/strategy/BoardCellsCreator.kt
@@ -1,4 +1,8 @@
-package tdd.minesweeper.domain
+package tdd.minesweeper.domain.strategy
+
+import tdd.minesweeper.domain.Area
+import tdd.minesweeper.domain.Cells
+import tdd.minesweeper.domain.Location
 
 interface BoardCellsCreator {
     fun createCells(
